cambios

+++ Framework/Libraies/ResultShoppingCartProduct.cs
@@ -16,7 +16,7 @@ namespace Framework.Libraies
         public string IdUser { get; set; }
         public Nullable<int> proId { get; set; }
         public Nullable<int> quantity { get; set; }
-        public Nullable<decimal> totalpriceprod { get; set; }
+        public Nullable<double> totalpriceprod { get; set; }
         public string cartid { get; set; }
     }
 }
